agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
langchain==0.1.0
langchain-openai==0.0.2
playwright==1.40.0
requests==2.31.0
python-multipart==0.0.6
asyncio-mqtt==0.16.1